+++ backend/src/main/java/com/trendfony/backend/product/controller/response/ProductInfoResponse.java
@@ -20,7 +20,7 @@ public class ProductInfoResponse {
     private Long categoryRank;
     private Long price;
     private Long reviewCnt;
-    private Long rating;
+    private float rating;
     private int isSoldout;
     private String storeName;
     private LocalDateTime createdAt;
