+++ backend/src/main/java/com/trendfony/backend/product/view/dto/CreateProductDto.java
@@ -17,7 +17,7 @@ public class CreateProductDto {
     private Long categoryRank;
     private Long price;
     private Long reviewCnt;
-    private Long rating;
+    private float rating;
     private int isSoldout;
     private String storeName;
 }
